roadmap画面で STEP1.5など 小数点を登録できるように修正。

+++ content/models.py
@@ -107,8 +107,8 @@ class Roadmap(models.Model):
     name = models.CharField(max_length=50)
     slug = models.SlugField(unique=True, blank=True, null=False)
 
-    # 表示順（STEP01, STEP02... の順番制御用）
-    order = models.PositiveIntegerField(default=1, db_index=True)
+    # 表示順（STEP1, STEP2... の順番制御用）
+    order = models.DecimalField(max_digits=4, decimal_places=1, default=1, db_index=True)
 
     # アコーディオン見出し下の短い説明
     summary = models.CharField(max_length=200, blank=True)
